fix: 코멘트 신고 로직 ReportController 및 ReportService로 이동
