+++ src/main/java/mocacong/server/repository/CommentReportRepository.java
@@ -1,12 +1,12 @@
 package mocacong.server.repository;
 
-import mocacong.server.domain.CommentReport;
+import mocacong.server.domain.Report;
 import mocacong.server.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentReportRepository extends JpaRepository<CommentReport, Long> {
+public interface CommentReportRepository extends JpaRepository<Report, Long> {
 
-    List<CommentReport> findAllByReporter(Member reporter);
+    List<Report> findAllByReporter(Member reporter);
 }
