+++ src/main/java/mocacong/server/repository/CommentRepository.java
@@ -5,11 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-
-    List<Comment> findAllByMemberId(Long memberId);
 
     Slice<Comment> findByMemberId(Long memberId, Pageable pageRequest);
